feat: Render contents nav and add event to contents nav
