+++ index.html
@@ -86,19 +86,8 @@
           <h3 class="contents-title">요일 연재 TOP <span>(1,620)</span></h3>
           <a href="#" class="pseudo-img">더보기</a>
         </div>
-        <nav class="contents-nav">
-          <ul>
-            <li>월</li>
-            <li class="selected">화</li>
-            <li>수</li>
-            <li>목</li>
-            <li>금</li>
-            <li>토</li>
-            <li>일</li>
-            <li>완결</li>
-          </ul>
-        </nav>
-        <ol class="contents-list small">
+        <nav class="contents-nav"></nav>
+        <ul class="contents-list small">
           <li class="list-item">
             <a href="#">
               <div class="img-container">
@@ -229,7 +218,7 @@
               <span class="reader pseudo-img">122.6만명</span>
             </a>
           </li>
-        </ol>
+        </ul>
       </section>
       <section class="contents anticipated-work">
         <div class="contents-header">
@@ -831,6 +820,6 @@
       <p>저작권 ok</p>
     </div>
   </footer>
-  <script src="src/render.js" type="module"></script>
+  <script src="src/index.js" type="module"></script>
 </body>
 </html>
