+++ src/components/index.js
@@ -1,5 +1,6 @@
 import Nav from './nav/Nav.js';
 import MainNav from './nav/MainNav.js';
+import ContentsNav from './nav/ContentsNav.js';
 import MainContents from './MainContents.js';
 
-export { Nav, MainNav, MainContents };
+export { Nav, MainNav, ContentsNav, MainContents };
