+++ src/components/nav/ContentsNav.js
@@ -1,0 +1,5 @@
+import { Nav } from '../index.js';
+
+const ContentsNav = (contentsNavItems) => `<nav class="contents-nav">${Nav(contentsNavItems)}</nav>`;
+
+export default ContentsNav;
