+++ src/data/index.js
@@ -1,40 +1,3 @@
-const MainNavItems = [
-  {
-    name: '홈',
-    isSelected: true,
-  },
-  {
-    name: '요일연재',
-    isSelected: false,
-  },
-  {
-    name: '웹툰',
-    isSelected: false,
-  },
-  {
-    name: '소년',
-    isSelected: false,
-  },
-  {
-    name: '드라마',
-    isSelected: false,
-  },
-  {
-    name: '로맨스',
-    isSelected: false,
-  },
-  {
-    name: '로판',
-    isSelected: false,
-  },
-  {
-    name: '액션무협',
-    isSelected: false,
-  },
-  {
-    name: 'BL',
-    isSelected: false,
-  },
-];
+import { MainNavItems, ContentsNavItems } from './navItems.js';
 
-export { MainNavItems };
+export { MainNavItems, ContentsNavItems };
