+++ src/data/navItems.js
@@ -1,0 +1,75 @@
+const MainNavItems = [
+  {
+    name: '홈',
+    isSelected: true,
+  },
+  {
+    name: '요일연재',
+    isSelected: false,
+  },
+  {
+    name: '웹툰',
+    isSelected: false,
+  },
+  {
+    name: '소년',
+    isSelected: false,
+  },
+  {
+    name: '드라마',
+    isSelected: false,
+  },
+  {
+    name: '로맨스',
+    isSelected: false,
+  },
+  {
+    name: '로판',
+    isSelected: false,
+  },
+  {
+    name: '액션무협',
+    isSelected: false,
+  },
+  {
+    name: 'BL',
+    isSelected: false,
+  },
+];
+
+const ContentsNavItems = [
+  {
+    name: '월',
+    isSelected: true,
+  },
+  {
+    name: '화',
+    isSelected: false,
+  },
+  {
+    name: '수',
+    isSelected: false,
+  },
+  {
+    name: '목',
+    isSelected: false,
+  },
+  {
+    name: '금',
+    isSelected: false,
+  },
+  {
+    name: '토',
+    isSelected: false,
+  },
+  {
+    name: '일',
+    isSelected: false,
+  },
+  {
+    name: '완결',
+    isSelected: false,
+  },
+]
+
+export { MainNavItems, ContentsNavItems };
